// src/features/svg-canvas/types/CoordinateTypes.ts
// 座標系に関する型定義

/**
 * 座標（x,y）の型定義
 */
export type Point = {
	x: number;
	y: number;
};

/**
 * 矩形の枠の型定義
 */
export type Box = {
	top: number;
	left: number;
	right: number;
	bottom: number;
	center: Point;
	leftTop: Point;
	leftBottom: Point;
	rightTop: Point;
	rightBottom: Point;
};

/**
 * 矩形の頂点の型定義
 */
export type RectangleVertices = {
	leftTopPoint: Point;
	leftBottomPoint: Point;
	rightTopPoint: Point;
	rightBottomPoint: Point;
	topCenterPoint: Point;
	leftCenterPoint: Point;
	rightCenterPoint: Point;
	bottomCenterPoint: Point;
};

/**
 * 楕円の頂点の型定義
 */
export type EllipseVertices = {
	leftTopPoint: Point;
	leftBottomPoint: Point;
	rightTopPoint: Point;
	rightBottomPoint: Point;
	topCenterPoint: Point;
	leftCenterPoint: Point;
	rightCenterPoint: Point;
	bottomCenterPoint: Point;
};